backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6
supabase>=2.0.0
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
PyPDF2>=3.0.0
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.0
numpy>=1.24.0
requests>=2.30.0
python-dotenv>=1.0.0
ollama>=0.1.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
aiohttp>=3.8.0

redis>=5.0.0
rq>=1.15.1
python-jose[cryptography]>=3.3.0
pydantic-settings>=2.2.1
pydantic>=2.0.0
websockets>=12.0
slowapi>=0.1.9
python-json-logger>=2.0.7
sentry-sdk[fastapi]>=1.40.0
